requirements.txt
google-cloud-bigquery>=3.0.0
google-adk
google-cloud-secret-manager
google-auth
black
flake8
isort
requests
beautifulsoup4
google-cloud-language
streamlit
pandas
streamlit
pandas
beautifulsoup4

